1679: Move screenshotting to separate context
